requirements.txt
requests>=2.31
pandas>=2.0
pydantic>=2.0
python-dateutil>=2.8
python-dotenv>=1.0
